Renames to_struct to to_struct!
This better fits with what it actually does I think. While it might not
actually raise (it depends how the casting functions and the accessors
are implemented) it gets across the fact that we don't stop. Although
perhaps we could call the other function to_struct_while. That may get
across that we are gonna bail out early in case of errors... Hmm.

Anyway for now we have gone with to_struct!. May change