Adds a first stab at to_struct
Needs testing more thoroughly but this implements a way for us to return
error tuples from casting functions and have the to_struct process stop
with an error. This could be handy if we want to do something off the
back of an error, maybe yell at the user or whatever.