More tests for to_struct
Enforces that cast functions return okay tuples.
Adds documentation
Adds guard for map accessor
We need to remove to_struct and add some tests for when there are
options on the fields.
We also need to have a better error object and we need to implement
enforcing not null on struct creation.

Then there is an open question about validations but I don't think we
need to re-invent ecto completely.