Adds has_many
Right. We add a has_many as a new field type to allow for the
collect_errors option. Essentially if nested schemas are implemented by
manually implementing a cast function that calls to_struct you can end
up in a situation where one call to to_struct collects errors and one
doesn't. It also means the decision as to whether you collect errors is
baked in at compile time (when you create the schema) for relations.
This is perhaps not ideal as I can imagine wanting the option at runtime
of doing either.

To get around this we need to make to_struct recursive which allows for
to_struct to control the options and apply them to all branches of the
to_struct function.

You could still implement a relation as a list_of schemas and implement
cast for those schemas but then you have to figure out for yourself what
to do when that call errors and you have to realise that the options you
supply for that field are baked in at compile time. I think this is the
best balance of not adding too much complexity and helping people not
shoot themselves in the foot.