Characteristcally we failed at git and did a lot.
1. Changed how aggregate works

We realised that as was there was no way of specifying different access
patterns for different aggregated fields. Like if you wanted one of the
keys to have a list of values that wasn't possible. So now we remove
aggregate from the access behaviour and make aggregated fields return a
map of values defined by a nested schema. This can be inline or not, but
essentially means we can get quite powerful transformations. The map
that gets created gets passed to the casting function for the aggregated
field.

We also added has_many I don't know if we did that already or not. Then
we added some more better schema validations on create including some
nice error messages when we raise compiler errors.

We add a to do list to the repo and refactor to_struct a bit to try and
make it manageable.