fixes the has_one, need to document the path rules though.
The rules are has_many goes to the list + 1
has_one needs to also be like a top level schema and have the outer
node's name there. This is possibly less flexible but more consistent.
Like you can't drop the same schema into a different place... so could
be more verbose.