Start experiment with Saxy XML parsing
This will be an experiment where we try to create a simple form
representation of the XML but only for nodes and attributes that exist
in a given schema. The idea is that this should be much faster and
should use less memory than some other approaches. Once working we can
benchmark against various other approaches.

This commit kicks us off with very basic skipping abilities. We need to
add more tests to check it all works as expected.