Fixes bug in to_struct
When we refactored we made it so that you would always call the casting
function with the empty value for has one and has many, which broke some
things.

Now the logic is:

If the value is not empty then we will call the cast fn with that non
empty value.

If it's empty it will either error (if it's not optional) or it will
return the empty value for the field if it is optional.